Error handling for initial data load and main data query
